libs/mcp/pyproject.toml
[project]
name = "sdif-mcp"
version = "0.2.5"
description = "SDIF MCP"
authors = [
    {name = "Syncpulse"}
]
maintainers = [
    {name = "Bryan Djafer", email = "bryan.djafer@syncpulse.fr"}
]
license = "MIT"
readme = "README.md"

requires-python = ">=3.10,<4.0"

[tool.poetry.dependencies]
fastmcp = ">=2.3.0,<3.0.0"
sdif-db = ">=0.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
sdif-mcp = "satif.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
pytest-cov = "^6.1.1"
pandas-stubs = "^2.2.0"
mypy = "^1.15.0"
ruff = "^0.11.10"
sdif-db = {path = "../sdif", develop = true}

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
    "D100",
    "D103",
    "D102",
    "D101",
    "D104",
    "D106",
    "D107",
    "D205",
    "D212",
    "D401",
    "T201",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
